Fixed télé-expertise app - added missing features & modern UI Complete consultation workflow, cost calculator, expertise requests working

+++ src/main/java/com/example/platform_tele_expertise_medicale/controller/ExpertiseServlet.java
@@ -141,7 +141,7 @@ public class ExpertiseServlet extends HttpServlet {
             throws ServletException, IOException {
         try {
             Long consultationId = Long.parseLong(request.getParameter("consultationId"));
-            Long specialiteId = Long.parseLong(request.getParameter("specialiteId"));
+            Integer specialiteId = Integer.parseInt(request.getParameter("specialiteId"));
             String question = request.getParameter("question");
             Priorite priorite = Priorite.valueOf(request.getParameter("priorite"));
